Move tests and use fixtures
